+++ lib/bot.js
@@ -18,7 +18,7 @@ class Bot {
     this.mode = mode
     this.dbCredentials = dbCredentials
     this.bot = new TelegramBot(this.getBaseConf())
-    require('./lib/index')(this.bot)
+    require('./index')(this.bot)
   }
 
   /**
